deser_incomplete::random_trailer
Trait RandomTrailer

pub trait RandomTrailer {
    // Required methods
    fn prepare_string_with_tag(&self, input: &mut String, tag: &str);
    fn prepare_vec_with_tag(&self, input: &mut Vec<u8>, tag: &str);
    fn remove_trailer(
        &self,
        string_like: &mut impl StringLike,
        random_tag: &str,
    ) -> bool;
}

Available on crate feature rand only.

For some data formats, we like to append some randomized trailer to the input before deserializing.

For JSON, this lets us support incomplete strings.

For YAML, we need this otherwise any incomplete string seems to cause deserialization to fail.

However, such a trailer then needs to be removed from any decoded strings.

Required Methods

fn prepare_string_with_tag(&self, input: &mut String, tag: &str)

Given a random tag, add the corresponding trailer to the input.

fn prepare_vec_with_tag(&self, input: &mut Vec<u8>, tag: &str)

Given a random tag, add the corresponding trailer to the input.

fn remove_trailer( &self, string_like: &mut impl StringLike, random_tag: &str, ) -> bool

Given a parsed string or bytes, detect whether the random tag was parsed, then take off the suffix of the parsed string that may not have been present in the original input.

Return whether the tag was detected. This simultaneously represents that we have encountered the end of the input, and we should probably stop trying to deserialize more.

This method should only modify the input if the tag is detected.

This method will not be called when parsing was done without a random tag.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.