deser_incomplete
Module source

Trait Source describes a resource that can be deserialized multiple times: a Deserializer together with the data it works on, such as a string reference or byte slice.

Structs

JsonBytesserde_json

Essentially serde_json::from_slice.

JsonStrserde_json

Essentially serde_json::from_str.

YamlBytesserde_yaml

Essentially serde_yaml::from_slice.

YamlStrserde_yaml

Essentially serde_yaml::from_str.

Traits

Source

Describes a resource that can be deserialized multiple times: a Deserializer together with the data it works on, such as a string reference or byte slice.