deser_incomplete::unstable
Struct UnstableCustomBehavior

#[non_exhaustive]
pub struct UnstableCustomBehavior {

    pub tolerate_deserializer_fail_after_visit_success: bool,
    pub fallback_any_as_none: bool,
    pub fallback_ignored_any_as_none: bool,
    pub fallback_default_bool: Option<bool>,
    pub fallback_int_zero: bool,
    pub fallback_default_float: Option<f32>,
    pub fallback_default_char: Option<char>,
    pub fallback_default_str: Option<&'static str>,
    pub fallback_bytes_empty: bool,
    pub fallback_none: bool,
    pub fallback_none_at_mandatory: bool,
    pub fallback_unit: bool,
    pub fallback_unit_at_mandatory: bool,
    pub fallback_unit_struct: bool,
    pub fallback_unit_struct_at_mandatory: bool,
    pub fallback_seq_empty: bool,
    pub fallback_seq_empty_at_root: bool,
    pub backtrack_seq_empty_for_value: bool,
    pub fallback_seq_skip_item: bool,
    pub backtrack_seq_skip_item: bool,
    pub fallback_tuple_empty: bool,
    pub fallback_tuple_skip_item: bool,
    pub backtrack_tuple_skip_item: bool,
    pub fallback_tuple_struct_empty: bool,
    pub fallback_tuple_struct_skip_item: bool,
    pub backtrack_tuple_struct_skip_item: bool,
    pub fallback_map_empty: bool,
    pub fallback_map_empty_at_root: bool,
    pub backtrack_map_empty_for_value: bool,
    pub fallback_map_skip_item: bool,
    pub backtrack_map_skip_item: bool,
    pub fallback_struct_empty: bool,
    pub fallback_struct_empty_at_root: bool,
    pub backtrack_struct_empty_for_value: bool,
    pub fallback_struct_skip_field: bool,
    pub backtrack_struct_skip_field: bool,
    pub fallback_unit_variant: bool,
    pub fallback_other_skip_item: bool,
    pub backtrack_other_skip_item: bool,
    pub allow_incomplete_string_in_key_or_variant: bool,

}

Available on crate feature unstable only.

Customize behavior.

This both controls

when deserializer encounters end-of-input, but we still have a chance to fill in the value and succeed deserialization, then we can make an educated guess based on what the data type expected (which method of crate::Deserializer was called).

For instance, when deserializing an option, JSON fa will choose the Some case, the deserializer will error, but can save deserialization and fill in none.

which points are eligible for backtracking, e.g. is it okay to retry but omit a list item, a map item, or convert Some into None.

The default is an educated guess that should make many data types deserialize successfully on the first backtrack, without trying to apply too many fallbacks: incomplete list items are skipped, but no fallback numbers are filled in.

Limitations

Consider that fallback values are not a good substitute for #[serde(default)]:

struct Point {
    // note: missing #[serde(default)]
    x: i32,
    // note: missing #[serde(default)]
    y: i32,
}

We could put a fallback of 0 for integers, and that will make {"x": 1, "y": parse successfully, but {"x": 1 will still not parse because all Points require a value for y.

Stability

This interface is not subject to semver (as it is unstable), and may change or be removed at any time.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

tolerate_deserializer_fail_after_visit_success: bool

If the visitor returns Ok but the deserializer returns an error, then return the visitor’s value.

fallback_any_as_none: boolfallback_ignored_any_as_none: boolfallback_default_bool: Option<bool>fallback_int_zero: boolfallback_default_float: Option<f32>fallback_default_char: Option<char>fallback_default_str: Option<&'static str>fallback_bytes_empty: boolfallback_none: bool

In case Deserializer::deserialize_option does not do anything, then just go in and visit a None.

fallback_none_at_mandatory: boolfallback_unit: bool

In case Deserializer::deserialize_unit does not do anything, then just go in and visit a unit.

Caveat

Especially with a random trailer, this can sometimes generate spurious list elements.

fallback_unit_at_mandatory: boolfallback_unit_struct: bool

In case Deserializer::deserialize_unit_struct does not do anything, then just go in and visit a unit struct.

Caveat

Especially with a random trailer, this can sometimes generate spurious list elements.

fallback_unit_struct_at_mandatory: boolfallback_seq_empty: boolfallback_seq_empty_at_root: boolbacktrack_seq_empty_for_value: boolfallback_seq_skip_item: bool

fallback_*_skip_item are probably a bad idea, because they can mask a backtracking point that is better.

backtrack_seq_skip_item: boolfallback_tuple_empty: boolfallback_tuple_skip_item: bool

fallback_*_skip_item are probably a bad idea, because they can mask a backtracking point that is better.

backtrack_tuple_skip_item: boolfallback_tuple_struct_empty: boolfallback_tuple_struct_skip_item: bool

fallback_*_skip_item are probably a bad idea, because they can mask a backtracking point that is better.

backtrack_tuple_struct_skip_item: boolfallback_map_empty: boolfallback_map_empty_at_root: boolbacktrack_map_empty_for_value: boolfallback_map_skip_item: bool

fallback_*_skip_item are probably a bad idea, because they can mask a backtracking point that is better.

backtrack_map_skip_item: boolfallback_struct_empty: boolfallback_struct_empty_at_root: boolbacktrack_struct_empty_for_value: boolfallback_struct_skip_field: bool

fallback_*_skip_item are probably a bad idea, because they can mask a backtracking point that is better.

backtrack_struct_skip_field: boolfallback_unit_variant: boolfallback_other_skip_item: bool

Whether it’s okay to fallback to skipping an element or field in circumstances other than Deserializer::deserialize_seq or Deserializer::deserialize_tuple.

fallback_*_skip_item are probably a bad idea, because they can mask a backtracking point that is better.

backtrack_other_skip_item: boolallow_incomplete_string_in_key_or_variant: bool

Whether incomplete strings should be allowed or rejected in MapAccess::next_key or EnumAccess::variant.

Implementations

impl UnstableCustomBehavior

pub fn no_fallbacks(self) -> Self

pub fn strict() -> Self

Maximally strict behavior. Probably this library behaves like ordinary deserialization with this behavior.

pub fn lenient() -> Self