deser_incomplete::unstable
Trait Reporter

pub trait Reporter: Clone {

    // Required methods
    fn report_deserialize_start_any(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_bool(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_i8(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_i16(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_i32(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_i64(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_i128(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_u8(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_u16(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_u32(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_u64(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_u128(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_f32(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_f64(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_char(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_str(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_string(
        &mut self,
        args: impl DeserializeStartArgs,
    );
    fn report_deserialize_start_bytes(
        &mut self,
        args: impl DeserializeStartArgs,
    );
    fn report_deserialize_start_byte_buf(
        &mut self,
        args: impl DeserializeStartArgs,
    );
    fn report_deserialize_start_option(
        &mut self,
        args: impl DeserializeStartArgs,
    );
    fn report_deserialize_start_unit(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_unit_struct(
        &mut self,
        args: impl DeserializeStartArgs,
        name: &'static str,
    );
    fn report_deserialize_start_newtype_struct(
        &mut self,
        args: impl DeserializeStartArgs,
        name: &'static str,
    );
    fn report_deserialize_start_seq(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_tuple(
        &mut self,
        args: impl DeserializeStartArgs,
        len: usize,
    );
    fn report_deserialize_start_tuple_struct(
        &mut self,
        args: impl DeserializeStartArgs,
        name: &'static str,
        len: usize,
    );
    fn report_deserialize_start_map(&mut self, args: impl DeserializeStartArgs);
    fn report_deserialize_start_struct(
        &mut self,
        args: impl DeserializeStartArgs,
        name: &'static str,
        fields: &'static [&'static str],
    );
    fn report_deserialize_start_enum(
        &mut self,
        args: impl DeserializeStartArgs,
        name: &'static str,
        variants: &'static [&'static str],
    );
    fn report_deserialize_start_identifier(
        &mut self,
        args: impl DeserializeStartArgs,
    );
    fn report_deserialize_start_ignored_any(
        &mut self,
        args: impl DeserializeStartArgs,
    );
    fn report_deserialize_finish(&mut self, error: Option<&dyn StdError>);
    fn report_deserialize_fallback_use_saved_value(&mut self);
    fn report_recv_visit_finish_primitive(
        &mut self,
        error: Option<&dyn StdError>,
    );
    fn report_recv_visit_start_bool(&mut self, v: bool);
    fn report_recv_visit_start_i8(&mut self, v: i8);
    fn report_recv_visit_start_i16(&mut self, v: i16);
    fn report_recv_visit_start_i32(&mut self, v: i32);
    fn report_recv_visit_start_i64(&mut self, v: i64);
    fn report_recv_visit_start_i128(&mut self, v: i128);
    fn report_recv_visit_start_u8(&mut self, v: u8);
    fn report_recv_visit_start_u16(&mut self, v: u16);
    fn report_recv_visit_start_u32(&mut self, v: u32);
    fn report_recv_visit_start_u64(&mut self, v: u64);
    fn report_recv_visit_start_u128(&mut self, v: u128);
    fn report_recv_visit_start_f32(&mut self, v: f32);
    fn report_recv_visit_start_f64(&mut self, v: f64);
    fn report_recv_visit_start_char(&mut self, v: char);
    fn report_recv_visit_start_str(&mut self, v: &str);
    fn report_recv_visit_start_borrowed_str(&mut self, v: &str);
    fn report_recv_visit_start_string(&mut self, v: &str);
    fn report_recv_visit_start_bytes(&mut self, v: &[u8]);
    fn report_recv_visit_start_borrowed_bytes(&mut self, v: &[u8]);
    fn report_recv_visit_start_byte_buf(&mut self, v: &[u8]);
    fn report_recv_visit_start_none(&mut self);
    fn report_recv_visit_start_some(&mut self);
    fn report_recv_visit_finish_some(&mut self, error: Option<&dyn StdError>);
    fn report_recv_visit_start_unit(&mut self);
    fn report_recv_visit_start_newtype_struct(&mut self);
    fn report_recv_visit_finish_newtype_struct(
        &mut self,
        error: Option<&dyn StdError>,
    );
    fn report_recv_visit_start_seq(&mut self);
    fn report_recv_visit_finish_seq(&mut self, error: Option<&dyn StdError>);
    fn report_recv_visit_start_map(&mut self);
    fn report_recv_visit_finish_map(&mut self, error: Option<&dyn StdError>);
    fn report_recv_visit_start_enum(&mut self);
    fn report_recv_visit_finish_enum(&mut self, error: Option<&dyn StdError>);
    fn report_new_halting_point(&mut self, point: &HaltingPoint);
    fn report_start_intervention(
        &mut self,
        reason: impl Debug,
        candidate_halting_point_for_next_attempt: Option<&HaltingPoint>,
        halting_point_stack: &[HaltingPoint],
    );
    fn report_start_fallback(&mut self);
    fn report_no_fallback(&mut self);
    fn report_fallback(&mut self, error: Option<&dyn StdError>);
    fn report_fallback_no_element(&mut self);
    fn report_seq_next_element_start(&mut self);
    fn report_seq_next_element_finish(
        &mut self,
        present: bool,
        error: Option<&dyn StdError>,
    );
    fn report_seq_next_element_skip(&mut self);
    fn report_map_next_key_start(&mut self);
    fn report_map_next_key_finish(
        &mut self,
        present: bool,
        error: Option<&dyn StdError>,
    );
    fn report_map_next_key_skip(&mut self);
    fn report_map_next_value_start(&mut self);
    fn report_map_next_value_finish(&mut self, error: Option<&dyn StdError>);
    fn report_access_past_end(&mut self);
    fn report_enum_start(&mut self);
    fn report_enum_finish(&mut self, error: Option<&dyn StdError>);
    fn report_variant_start_unit_variant(&mut self);
    fn report_variant_start_newtype_variant(&mut self);
    fn report_variant_start_tuple_variant(&mut self, len: usize);
    fn report_variant_start_struct_variant(
        &mut self,
        fields: &'static [&'static str],
    );
    fn report_variant_finish(&mut self, error: Option<&dyn StdError>);
    fn report_allow_incomplete_string(&mut self);
    fn report_reject_incomplete_string(&mut self);

}

Available on crate feature unstable only.

Reporters are for inspecting the normal behavior of this library.

Logic errors are typically not exposed here, and will be logged with tracing instead.

Required Methods

fn report_deserialize_start_any(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_bool(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_i8(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_i16(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_i32(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_i64(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_i128(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_u8(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_u16(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_u32(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_u64(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_u128(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_f32(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_f64(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_char(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_str(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_string(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_bytes(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_byte_buf(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_option(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_unit(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_unit_struct( &mut self, args: impl DeserializeStartArgs, name: &'static str, )

fn report_deserialize_start_newtype_struct( &mut self, args: impl DeserializeStartArgs, name: &'static str, )

fn report_deserialize_start_seq(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_tuple( &mut self, args: impl DeserializeStartArgs, len: usize, )

fn report_deserialize_start_tuple_struct( &mut self, args: impl DeserializeStartArgs, name: &'static str, len: usize, )

fn report_deserialize_start_map(&mut self, args: impl DeserializeStartArgs)

fn report_deserialize_start_struct( &mut self, args: impl DeserializeStartArgs, name: &'static str, fields: &'static [&'static str], )

fn report_deserialize_start_enum( &mut self, args: impl DeserializeStartArgs, name: &'static str, variants: &'static [&'static str], )

fn report_deserialize_start_identifier( &mut self, args: impl DeserializeStartArgs, )

fn report_deserialize_start_ignored_any( &mut self, args: impl DeserializeStartArgs, )

fn report_deserialize_finish(&mut self, error: Option<&dyn StdError>)

fn report_deserialize_fallback_use_saved_value(&mut self)

fn report_recv_visit_finish_primitive(&mut self, error: Option<&dyn StdError>)

This is called after visiting anything that doesn’t have its own report_end_* method.

fn report_recv_visit_start_bool(&mut self, v: bool)

fn report_recv_visit_start_i8(&mut self, v: i8)

fn report_recv_visit_start_i16(&mut self, v: i16)

fn report_recv_visit_start_i32(&mut self, v: i32)

fn report_recv_visit_start_i64(&mut self, v: i64)

fn report_recv_visit_start_i128(&mut self, v: i128)

fn report_recv_visit_start_u8(&mut self, v: u8)

fn report_recv_visit_start_u16(&mut self, v: u16)

fn report_recv_visit_start_u32(&mut self, v: u32)

fn report_recv_visit_start_u64(&mut self, v: u64)

fn report_recv_visit_start_u128(&mut self, v: u128)

fn report_recv_visit_start_f32(&mut self, v: f32)

fn report_recv_visit_start_f64(&mut self, v: f64)

fn report_recv_visit_start_char(&mut self, v: char)

fn report_recv_visit_start_str(&mut self, v: &str)

fn report_recv_visit_start_borrowed_str(&mut self, v: &str)

fn report_recv_visit_start_string(&mut self, v: &str)

fn report_recv_visit_start_bytes(&mut self, v: &[u8])

fn report_recv_visit_start_borrowed_bytes(&mut self, v: &[u8])

fn report_recv_visit_start_byte_buf(&mut self, v: &[u8])

fn report_recv_visit_start_none(&mut self)

fn report_recv_visit_start_some(&mut self)

fn report_recv_visit_finish_some(&mut self, error: Option<&dyn StdError>)

fn report_recv_visit_start_unit(&mut self)

fn report_recv_visit_start_newtype_struct(&mut self)

fn report_recv_visit_finish_newtype_struct( &mut self, error: Option<&dyn StdError>, )

fn report_recv_visit_start_seq(&mut self)

fn report_recv_visit_finish_seq(&mut self, error: Option<&dyn StdError>)

fn report_recv_visit_start_map(&mut self)

fn report_recv_visit_finish_map(&mut self, error: Option<&dyn StdError>)

fn report_recv_visit_start_enum(&mut self)

fn report_recv_visit_finish_enum(&mut self, error: Option<&dyn StdError>)

fn report_new_halting_point(&mut self, point: &HaltingPoint)

fn report_start_intervention( &mut self, reason: impl Debug, candidate_halting_point_for_next_attempt: Option<&HaltingPoint>, halting_point_stack: &[HaltingPoint], )

fn report_start_fallback(&mut self)

The deserializer failed without consuming the visitor. We start computing one of the Fallbacks.

fn report_no_fallback(&mut self)

The deserializer failed without consuming the visitor, but we chose not to apply one of the Fallbacks. (This is reported when the fallback does not take the visitor.)

fn report_fallback(&mut self, error: Option<&dyn StdError>)

The deserializer failed without consuming the visitor, and one of the Fallbacks was applied, or at least attempted.

fn report_fallback_no_element(&mut self)

From an Access type, we return as a fallback that there is no element left.

fn report_seq_next_element_start(&mut self)

fn report_seq_next_element_finish( &mut self, present: bool, error: Option<&dyn StdError>, )

fn report_seq_next_element_skip(&mut self)

The next element was requested from SeqAccess, but we’re not checking if there is a next element, we just decide it it’s not going to be there.

This could be for instance because we encountered our halting point.

fn report_map_next_key_start(&mut self)

fn report_map_next_key_finish( &mut self, present: bool, error: Option<&dyn StdError>, )

fn report_map_next_key_skip(&mut self)

The next key was requested from MapAccess, but we’re not checking if there is a next field, we just decide it it’s not going to be there.

This could be for instance because we encountered our halting point.

fn report_map_next_value_start(&mut self)

fn report_map_next_value_finish(&mut self, error: Option<&dyn StdError>)

fn report_access_past_end(&mut self)

The data type attempted to read a value from a collection after we already reported that there are no more.

fn report_enum_start(&mut self)

fn report_enum_finish(&mut self, error: Option<&dyn StdError>)

fn report_variant_start_unit_variant(&mut self)

fn report_variant_start_newtype_variant(&mut self)

fn report_variant_start_tuple_variant(&mut self, len: usize)

fn report_variant_start_struct_variant( &mut self, fields: &'static [&'static str], )

fn report_variant_finish(&mut self, error: Option<&dyn StdError>)

fn report_allow_incomplete_string(&mut self)

fn report_reject_incomplete_string(&mut self)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.